refactor similarity into matchers.py Move the similarity matcher into its own module. This is to accomodate new stuff that will be added
